requirements.txt
Streamlit==1.38.0
Pillow==10.4.0
pandas==2.2.3
requests==2.32.3